+++ rage/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+
+## [0.8.0] - 2022-05-02
 ### Changed
 - MSRV is now 1.56.0.
 - When both reading input from the terminal (e.g. if the user is typing the